+++ spec/openapi/parsers/ext/blueprinter_spec.rb
@@ -14,13 +14,133 @@ RSpec.describe Rage::OpenAPI::Parsers::Ext::Blueprinter do
     let_class("UserBlueprint") do
       <<~'RUBY'
         class UserBlueprint < Blueprinter::Base
+        end
+      RUBY
+    end
+    it do
+      is_expected.to eq({ "type" => "object" })
+    end
+  end
+
+  context "with basic fields" do
+    let_class("UserBlueprint") do
+      <<~'RUBY'
+        class UserBlueprint < Blueprinter::Base
           fields :id, :name, :email, :age
         end
       RUBY
     end
-
     it do
-      is_expected.to eq({ "type" => "object" })
+      is_expected.to eq({
+        "type" => "object",
+        "properties" => {
+          "id" => { "type" => "string" },
+          "name" => { "type" => "string" },
+          "email" => { "type" => "string" },
+          "age" => { "type" => "string" }
+        }
+      })
+    end
+  end
+
+  context "with identifier" do
+    let_class("UserBlueprint") do
+      <<~'RUBY'
+        class UserBlueprint < Blueprinter::Base
+          identifier :uuid
+        end
+      RUBY
+    end
+    it do
+      is_expected.to eq({
+        "type" => "object",
+        "properties" => {
+          "uuid" => { "type" => "string" }
+        }
+      })
+    end
+  end
+
+  context "with a single field" do
+    let_class("UserBlueprint") do
+      <<~'RUBY'
+        class UserBlueprint < Blueprinter::Base
+          field :email
+        end
+      RUBY
+    end
+    it do
+      is_expected.to eq({
+        "type" => "object",
+        "properties" => {
+          "email" => { "type" => "string" }
+        }
+      })
+    end
+  end
+
+  context "with field name alias" do
+    let_class("UserBlueprint") do
+      <<~'RUBY'
+        class UserBlueprint < Blueprinter::Base
+          field :email, name: :login
+        end
+      RUBY
+    end
+    it do
+      is_expected.to eq({
+        "type" => "object",
+        "properties" => {
+          "login" => { "type" => "string" }
+        }
+      })
+    end
+  end
+
+  context "with a block field" do
+    let_class("UserBlueprint") do
+      <<~'RUBY'
+        class UserBlueprint < Blueprinter::Base
+          field(:full_name) { |u| "#{u.first_name} #{u.last_name}" }
+        end
+      RUBY
+    end
+    it do
+      is_expected.to eq({
+        "type" => "object",
+        "properties" => {
+          "full_name" => { "type" => "string" }
+        }
+      })
+    end
+  end
+
+  context "with all declaration types combined" do
+    let_class("UserBlueprint") do
+      <<~'RUBY'
+        class UserBlueprint < Blueprinter::Base
+          identifier :uuid
+          fields :id, :name, :age
+          field :email, name: :login
+          fields :first_name, :last_name
+          field(:full_name) { |u| "#{u.first_name} #{u.last_name}" }
+        end
+      RUBY
+    end
+    it do
+      is_expected.to eq({
+        "type" => "object",
+        "properties" => {
+          "uuid" => { "type" => "string" },
+          "id" => { "type" => "string" },
+          "name" => { "type" => "string" },
+          "age" => { "type" => "string" },
+          "login" => { "type" => "string" },
+          "first_name" => { "type" => "string" },
+          "last_name" => { "type" => "string" },
+          "full_name" => { "type" => "string" }
+        }
+      })
     end
   end
 end